Cart repo Commit by rakesh

+++ ShoppingCart/DTOs/CartDto.cs
@@ -11,7 +11,7 @@ namespace ShoppingCart.DTOs
         public int UserId { get; set; }
         public int BookId { get; set; }
         public int? Quantity { get; set; }
-        public int? CartTotal { get; set; }
+        public decimal? CartTotal { get; set; }
         public decimal? DiscountPer { get; set; }
         public decimal NetPay { get; set; }
         public bool? IsActive { get; set; }
